+++ src/controllers/CacheController.php
@@ -8,14 +8,24 @@ namespace putyourlightson\blitz\controllers;
 use Craft;
 use craft\errors\MissingComponentException;
 use craft\web\Controller;
+use craft\web\View;
 use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\jobs\WarmCacheJob;
 use yii\base\Exception;
 use yii\web\BadRequestHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\web\Response;
 
 class CacheController extends Controller
 {
+    // Properties
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    protected $allowAnonymous = true;
+
     // Public Methods
     // =========================================================================
 
@@ -23,56 +33,42 @@ class CacheController extends Controller
      * Clears the cache.
      *
      * @return Response
-     * @throws BadRequestHttpException
-     * @throws MissingComponentException
      */
     public function actionClear(): Response
     {
         Blitz::$plugin->invalidate->clearCache(false);
 
-        Craft::$app->getSession()->setNotice(Craft::t('blitz', 'Blitz cache successfully cleared.'));
-
-        return $this->redirectToPostedUrl();
+        return $this->_getResponse('Blitz cache successfully cleared.');
     }
 
     /**
      * Flushes the cache.
      *
      * @return Response
-     * @throws BadRequestHttpException
-     * @throws MissingComponentException
      */
     public function actionFlush(): Response
     {
         Blitz::$plugin->invalidate->clearCache(true);
 
-        Craft::$app->getSession()->setNotice(Craft::t('blitz', 'Blitz cache successfully flushed.'));
-
-        return $this->redirectToPostedUrl();
+        return $this->_getResponse('Blitz cache successfully flushed.');
     }
 
     /**
      * Refreshes expired cache.
      *
      * @return Response
-     * @throws BadRequestHttpException
-     * @throws MissingComponentException
      */
     public function actionRefreshExpired(): Response
     {
         Blitz::$plugin->invalidate->refreshExpiredCache();
 
-        Craft::$app->getSession()->setNotice(Craft::t('blitz', 'Expired Blitz cache successfully refreshed.'));
-
-        return $this->redirectToPostedUrl();
+        return $this->_getResponse('Expired Blitz cache successfully refreshed.');
     }
 
     /**
      * Warms the cache.
      *
      * @return Response
-     * @throws BadRequestHttpException
-     * @throws MissingComponentException
      * @throws Exception
      */
     public function actionWarm(): Response
@@ -80,9 +76,7 @@ class CacheController extends Controller
         $settings = Blitz::$plugin->getSettings();
 
         if (!$settings->cachingEnabled) {
-            Craft::$app->getSession()->setError(Craft::t('blitz', 'Blitz caching is disabled.'));
-
-            return $this->redirectToPostedUrl();
+            return $this->_getResponse('Blitz caching is disabled.');
         }
 
         // Get URLs before flushing the cache
@@ -92,7 +86,59 @@ class CacheController extends Controller
 
         Craft::$app->getQueue()->push(new WarmCacheJob(['urls' => $urls]));
 
-        Craft::$app->getSession()->setNotice(Craft::t('blitz', 'Blitz cache successfully queued for warming.'));
+        return $this->_getResponse('Blitz cache successfully queued for warming.');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function beforeAction($action)
+    {
+        parent::beforeAction($action);
+
+        $request = Craft::$app->getRequest();
+
+        // Require permission if posted from utility
+        if ($request->getIsPost() && $request->getParam('utility')) {
+            $this->requirePermission('blitz:cache-utility');
+        }
+        else {
+            // Verify API key
+            $apiKey = $request->getParam('key');
+
+            $settings = Blitz::$plugin->getSettings();
+
+            if (empty($apiKey) || empty($settings->apiKey) || $apiKey != Craft::parseEnv($settings->apiKey)) {
+                throw new ForbiddenHttpException('Unauthorised access.');
+            }
+        }
+
+        return true;
+    }
+
+    // Private Methods
+    // =========================================================================
+
+    /**
+     * Returns a response.
+     *
+     * @param string $message
+     *
+     * @return Response
+     */
+    private function _getResponse(string $message)
+    {
+        $request = Craft::$app->getRequest();
+
+        // If front-end site or JSON request
+        if (Craft::$app->getView()->templateMode == View::TEMPLATE_MODE_SITE || $request->getAcceptsJson()) {
+            return $this->asJson([
+                'success' => true,
+                'message' => $message,
+            ]);
+        }
+
+        Craft::$app->getSession()->setNotice(Craft::t('blitz', $message));
 
         return $this->redirectToPostedUrl();
     }